refactor(auth): auth 도메인 디렉토리 리팩토링
