+++ src/main/java/me/jinjjahalgae/domain/auth/usecase/logout/LogoutUseCase.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.auth.usecase.interfaces;
+package me.jinjjahalgae.domain.auth.usecase.logout;
 
 public interface LogoutUseCase {
     void execute(Long userId);
