+++ src/main/java/me/jinjjahalgae/domain/auth/usecase/refresh/RefreshUseCase.java
@@ -1,0 +1,8 @@
+package me.jinjjahalgae.domain.auth.usecase.refresh;
+
+import me.jinjjahalgae.domain.auth.usecase.refresh.dto.RefreshRequest;
+import me.jinjjahalgae.domain.auth.usecase.refresh.dto.RefreshResponse;
+
+public interface RefreshUseCase {
+    RefreshResponse execute(RefreshRequest request);
+}
