+++ src/main/java/me/jinjjahalgae/presentation/api/docs/auth/refresh/RefreshSwaggerResponse.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.presentation.api.docs.auth.refresh;
 
-import me.jinjjahalgae.domain.auth.dto.refresh.RefreshResponse;
+import me.jinjjahalgae.domain.auth.usecase.refresh.dto.RefreshResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 
 public class RefreshSwaggerResponse extends CommonResponse<RefreshResponse> {
